Extended setup-ssh to set second variable as well.
I still need to change the path for the temp output file.
It currently outputs the file to the current path, i need
to change it so that it outputs to the directory of the cmd file.

I also still need to add the rest of the setup process,
like ssh-add and prompting for a pass phrase.